+++ app/src/main/java/io/github/protino/codewatch/remote/model/project/summary/Language.java
@@ -20,7 +20,7 @@ public class Language {
     private String name;
     @SerializedName("percent")
     @Expose
-    private Integer percent;
+    private Double percent;
     @SerializedName("seconds")
     @Expose
     private Integer seconds;
@@ -63,11 +63,11 @@ public class Language {
         this.name = name;
     }
 
-    public Integer getPercent() {
+    public Double getPercent() {
         return percent;
     }
 
-    public void setPercent(Integer percent) {
+    public void setPercent(Double percent) {
         this.percent = percent;
     }
